parse money into Decimal

+++ dnsimple/struct/charge.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 from dnsimple.struct import Struct
 
@@ -15,9 +16,9 @@ class Charge(Struct):
     def __init__(self, data):
         super().__init__(data)
         if self.total_amount is not None:
-            self.total_amount = float(self.total_amount)
+            self.total_amount = Decimal(self.total_amount)
         if self.balance_amount is not None:
-            self.balance_amount = float(self.balance_amount)
+            self.balance_amount = Decimal(self.balance_amount)
 
 
 @dataclass
@@ -31,4 +32,4 @@ class ChargeItem(Struct):
     def __init__(self, data):
         super().__init__(data)
         if self.amount is not None:
-            self.amount = float(self.amount)
+            self.amount = Decimal(self.amount)
